refactor(featues/partyroom/enter): 전반적으로 개선
- api 세그먼트와 lib 세그먼트 구분 명확히 함
- setup info, notice fetch를 서비스 인스턴스 그대로 가져다 쓰도록 설정 (캐시가 필요없는 1회성 api기에 query를 쓸 필요가 없음)
- 실행 시 PartyroomClient의 connect 리스너로 등록하는 행위를 슬라이스 내부에서 하도록 변경 (이전에는 app 레이어에서 수행했음)
