+++ src/entities/current-partyroom/index.ts
@@ -4,5 +4,3 @@ export * as Dj from './model/dj.model';
 export * as Playback from './model/playback.model';
 
 export { default as useHandlePartyroomSubscriptionEvent } from './lib/handle-subscription-event';
-
-export { default as fetchPartyroomSetUpInfo } from './api/fetch-set-up-info';
