+++ src/shared/api/http/query-keys.ts
@@ -8,6 +8,5 @@ export enum QueryKeys {
   PlaylistMusics = 'playlist/musics',
   Musics = 'musics',
   DjingQueue = 'partyroom/djing-queue',
-  Notice = 'partyroom/notice',
   Participants = 'partyroom/participants',
 }
